# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Authentication",
        "description": "API de gestion des utilisateurs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/admin/employe/create": {
            "post": {
                "tags": [
                    "Employé"
                ],
                "summary": "Créer un compte employé",
                "description": "Crée un nouveau compte employé et génère un mot de passe temporaire. Réservé aux administrateurs.",
                "operationId": "afcf21c21d3c80a9a2740860204c7604",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "employe@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employé créé avec succès. Le mot de passe temporaire est retourné (à envoyer par email en production).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Employé créé avec succès"
                                        },
                                        "password": {
                                            "description": "Mot de passe temporaire généré automatiquement",
                                            "type": "string",
                                            "example": "xK9#mP2!qL"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "500": {
                        "description": "Rôle employé introuvable en base"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/admin/employe/{id}": {
            "get": {
                "tags": [
                    "Employé"
                ],
                "summary": "Afficher un employé",
                "description": "Retourne les informations d'un employé par son identifiant. Réservé aux administrateurs.",
                "operationId": "4bc019814b4d240fd20ae58f859099a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de l'employé",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employé trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nom": {
                                            "type": "string",
                                            "example": "Dupont"
                                        },
                                        "prenom": {
                                            "type": "string",
                                            "example": "Marie"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "employe@example.com"
                                        },
                                        "statut": {
                                            "type": "string",
                                            "example": "actif"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "404": {
                        "description": "Employé introuvable"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Employé"
                ],
                "summary": "Supprimer un compte employé",
                "description": "Supprime un compte employé. Les comptes administrateurs ne peuvent pas être supprimés via cette route. Réservé aux administrateurs.",
                "operationId": "4ed70c54f1185fbfb259cbf87e7587e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de l'employé",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Compte supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès interdit — tentative de suppression d'un administrateur ou d'un non-employé"
                    },
                    "404": {
                        "description": "Employé introuvable"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/admin/employe": {
            "get": {
                "tags": [
                    "Employé"
                ],
                "summary": "Lister tous les employés",
                "description": "Retourne la liste de tous les comptes employés. Réservé aux administrateurs.",
                "operationId": "4688126dd70f2585dbf82023bc6b0cc3",
                "responses": {
                    "200": {
                        "description": "Liste des employés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nom": {
                                                "type": "string",
                                                "example": "Dupont"
                                            },
                                            "prenom": {
                                                "type": "string",
                                                "example": "Marie"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "employe@example.com"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "actif"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès interdit"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Employé"
                ],
                "summary": "Modifier plusieurs employés",
                "description": "Met à jour les informations d'un ou plusieurs employés en une seule requête. Chaque élément doit contenir un 'id'. Réservé aux administrateurs.",
                "operationId": "5087c17f0d476dd0533014ad31ff2a34",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "required": [
                                        "id"
                                    ],
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "employe@example.com"
                                        },
                                        "nom": {
                                            "type": "string",
                                            "example": "Dupont"
                                        },
                                        "prenom": {
                                            "type": "string",
                                            "example": "Marie"
                                        },
                                        "telephone": {
                                            "type": "string",
                                            "example": "0612345678"
                                        },
                                        "adresse": {
                                            "type": "string",
                                            "example": "5 rue de la Paix"
                                        },
                                        "ville": {
                                            "type": "string",
                                            "example": "Lyon"
                                        },
                                        "pays": {
                                            "type": "string",
                                            "example": "France"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employés mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Employés mis à jour"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "409": {
                        "description": "Email déjà utilisé"
                    },
                    "422": {
                        "description": "Erreurs de validation par index"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "name": "PHPSESSID",
                "in": "cookie"
            }
        }
    },
...
}